[DEL] : contactsTextField 삭제
